requirements.txt
pyspark==3.5.0
apache-airflow==2.8.0
pandas==2.1.0
numpy==1.26.0
pytest==7.4.0
boto3==1.34.0
pyarrow==14.0.0
python-dotenv==1.0.0
great-expectations==0.18.0
